pyproject.toml
[build-system]
requires = ["poetry==1.0.0"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "legal_basis_api"
description = "Legal Basis for Consent API"
version = "0.1.0"
readme = "README.md"
authors = []


[tool.poetry.dependencies]
python = "3.7.*"
django = "^2.2"
django-split-settings = "^1.0"
django-axes = "^5.0"
django-csp = "^3.5"
django-health-check = "^3.11"
django-http-referrer-policy = "^1.0"
django-feature-policy = "^3.1"
python-decouple = "^3.3"
dump-env = "^1.1"
bcrypt = "^3.1"
dj-database-url = "^0.5.0"
typing-extensions = "^3.7.4"
waitress = "^1.3.1"
whitenoise = {extras = ["brotli"], version = "^5.0"}
djangorestframework = "^3.10.3"
drf-yasg = {extras = ["validation"], version = "^1.17.0"}
psycopg2-binary = "^2.8.4"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^2.1"
django-querycount = "^0.7"
nplusone = "^1.0"

flake8-pytest = "^1.3"
flake8-pytest-style = "^0.1.3"
flake8-django = "^0.0.4"
xenon = "^0.5"

pytest = "^5.3"
pytest-django = "^3.7"
pytest-cov = "^2.8"
pytest-randomly = "^3.1"
pytest-deadfixtures = "^2.1"
pytest-testmon = "^1.0"
pytest-timeout = "^1.3"
django-test-migrations = "^0.1.0"
hypothesis = "^4.53"

mypy = "^0.750"
django-stubs = "^1.3"

ipython = "^7.10"

sphinx = "^2.2"
sphinx-autodoc-typehints = "^1.10"
doc8 = "^0.8"

yamllint = "^1.19"
safety = "^1.8"
dotenv-linter = "^0.1"
django-migration-linter = "^1.4"
django_extensions = "^2.2.5"
flake8-isort = "^2.8.0"
